Handle GrabKeyboard failure
Hopefully this is a reasonable way to go;
in any case either should only be able to fail
subject to a hopefully‐rare race condition,
esp. since active grabs normally ought to only be engaged
through active user request,
which would be tricky(!) to interleave here.

As of commit, this is also untested ngl

⟨再⟩ is the only unambiguous ∼cross‐linguistic representation
(besides the subtly distinct ⟨又⟩)
of ‘again’ ꝥ I could think of…