Document another foible
Because I can't currently be bothered reworking the whole event loop.
Would end up wanting a switch stmt in all likelihood,
and possibly a switch to more of a state‐machine structure,
and all for the relatively minimal benefit
of not having to remember to restart the dæmon
on the rare(?) occasion of futzing around w/ one's kb layout.

Reserving ofc the right to change my mind after all