Work around a bug in Xlib
To wit — Xlib tries to be ‘helpful’:
if XChangeKeyboardMapping(3) is called
to set a key to only one keysym,
but that keysym is one of a handful of casing pairs
(manifesting at least with common precomposed accented characters
such as ⟨é⟩ or ⟨ü⟩),
Xlib seems to think it appropriate to instead
set the key to produce the casing pair,
depending on shift state.
With the consequence
ꝥ if a compose sequence intended to produce a lowercase character
ends w/ a shifted key,
the indirection of emitting a new key event
causes clients to interpret the keycode with the shifted state
and emit a capital instead.

So we have to explicitly set 2 identical characters for the key.

As usual, ‘helpful’ extensions ꝥ conflict w/ naïve usage
turn out to cause problems. 😐︎

Fortunately zeroing a key still works w/ just 1 element.
So for resetting it we only have to add the pointer indirection